+++ server/src/main/java/client/MapleCharacter.java
@@ -41,11 +41,11 @@ import database.DatabaseException;
 import database.LoginService;
 import database.LoginState;
 import database.TeleportRockService;
-import handling.ServerMigration;
 import handling.cashshop.CashShop;
 import handling.channel.ChannelServer;
 import handling.channel.handler.utils.PartyHandlerUtils.PartyOperation;
 import handling.world.Broadcast;
+import handling.world.ServerMigration;
 import handling.world.WorldServer;
 import handling.world.buddy.MapleBuddyList;
 import handling.world.guild.GuildManager;
