+++ server/src/main/java/handling/channel/handler/EnterCashShopHandler.java
@@ -3,9 +3,9 @@ package handling.channel.handler;
 import client.MapleCharacter;
 import client.MapleClient;
 import constants.MapConstants;
-import handling.ServerMigration;
 import handling.cashshop.CashShopServer;
 import handling.channel.ChannelServer;
+import handling.world.ServerMigration;
 import handling.world.WorldServer;
 import handling.world.helper.CharacterTransfer;
 import handling.world.messenger.MapleMessengerCharacter;
