+++ server/src/main/java/handling/channel/handler/InterServerHandler.java
@@ -7,11 +7,11 @@ import client.MapleQuestStatus;
 import client.skill.SkillFactory;
 import constants.ServerConstants;
 import database.LoginState;
-import handling.ServerMigration;
 import handling.channel.ChannelServer;
 import handling.channel.CharacterIdChannelPair;
 import handling.channel.handler.utils.PartyHandlerUtils.PartyOperation;
 import handling.world.FindCommand;
+import handling.world.ServerMigration;
 import handling.world.WorldServer;
 import handling.world.alliance.AllianceManager;
 import handling.world.buddy.BuddyListEntry;
