+++ server/src/main/java/handling/world/WorldServer.java
@@ -2,8 +2,6 @@ package handling.world;
 
 import client.anticheat.CheaterData;
 import database.AccountData;
-import handling.MigrationService;
-import handling.MigrationServiceImpl;
 import handling.cashshop.CashShopServer;
 import handling.channel.ChannelServer;
 import handling.channel.PlayerStorage;
